requirements.txt
numpy>=1.24.0
sounddevice>=0.4.6
scipy>=1.10.0
deepgram-sdk
anthropic
google-genai